requirements.txt
numpy==2.1.2
pandas==2.2.3
pytz==2024.2
requests==2.32.3
python-dotenv==1.0.1